feat:data and util modules are connected
